GrpcMedData/requirements.txt
# requirements.txt

greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.29.0
psycopg2-binary==2.9.10
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
alembic==1.14.0
grpcio-reflection==1.68.0